CHORE: move poorly placed tests to the correct location
