+++ internal/test/issues/issue-1093/child.cfg.yaml
@@ -6,5 +6,5 @@ generate:
   strict-server: true
   models: true
 import-mapping:
-  parent.api.yaml: github.com/deepmap/oapi-codegen/examples/import-mapping/api/parent
+  parent.api.yaml: github.com/deepmap/oapi-codegen/internal/test/issues/issue-1093/api/parent
 output: ../../examples/import-mapping/api/child/child.gen.go
